- eliminate compiler warnings - use %g instead of %f: shorter string for sanity and no buffer overflow - Double: fix error in little-endian conversion - Bitstring: accept [{T,F...}] or B' (was broken from script) - encoded wildcards in time and date as 0xFE instead of 0xFF - BACnetPriorityValue: add type checks to avoid assert on null choice versus value choice

+++ VTS3/BACnetLibrary/BACnet.hpp
@@ -1230,7 +1230,7 @@ public:
 	}
 	void SetElement(unsigned int index, MyDataType* src)
 	{
-		if(index >= listSequenceOf.GetCount())
+		if((int)index >= listSequenceOf.GetCount())
 			throw "The specified index is beyond the scope of the sequenceof.";
 		POSITION pos = listSequenceOf.FindIndex(index);		
 		MyDataType* elem=listSequenceOf.GetAt(pos);
